refactor to be a bit more extensible
