+++ tests/tcp/HSLTCPTest.php
@@ -141,7 +141,7 @@ final class HSLTCPTest extends HackTest {
       // Portability:
       // - MacOS only requires SO_REUSEADDR to be set on the new socket
       // - Linux requires SO_REUSEADDR on both the old and the new socket
-      shape('SO_REUSEADDR' => true),
+      shape('socket_options' => shape('SO_REUSEADDR' => true)),
     );
     $port = $s1->getLocalAddress()[1];
     concurrent {
@@ -164,7 +164,7 @@ final class HSLTCPTest extends HackTest {
       IPProtocolVersion::IPV4,
       '127.0.0.1',
       $port,
-      shape('SO_REUSEADDR' => true),
+      shape('socket_options' => shape('SO_REUSEADDR' => true)),
     );
     concurrent {
       $client_recv = await async {
